moved rendering of plot border into gardenAPI and added border render while holding sprayonator
